+++ test/unit.js
@@ -13,7 +13,6 @@ async function test([filename, fileContent = ''], [type, testVal], opts) {
 		'files': actual.files.results[filename],
 		'size': actual.files.bytes,
 		'count': actual.files.count,
-		'alternatives_count': Object.entries(actual.files.alternatives).length,
 	}[type];
 	const result = testContent === testVal;
 	i = `${+i + 1}`.padStart(2, '0');
@@ -34,7 +33,6 @@ async function unitTest() {
 	await test(['x.cpp'], ['files', 'C++']);
 	await test(['x.c'], ['files', 'C']);
 	await test(['x.R'], ['files', 'R']);
-	await test(['.m'], ['alternatives_count', 1])
 	await test(['view.tsx'], ['files', 'TypeScript']);
 	await test(['view.tsx'], ['files', 'TSX'], { childLanguages: true });
 	desc('filenames');